requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
groq==0.8.0
django-filter==24.2
markdown==3.6
pygments==2.18.0
